+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Buffs replacing an armor bonus did not disable the enhancement armor bonus of the armor they just replaced
 - Hit points, wounds and vigor weren't limited by their maximum values
+- Racial and circumstance modifiers didn't stack
 
 ### Changelog